server.jsの分割
